// src/app/interfaces/contact.interface.ts
export interface IContactForm {
    name: string;
    email: string;
    check: boolean;
    recruiter: boolean;
    progress: boolean;
    subject: string;
}

export interface IContact {
    name: string;
    email: string;
    check: boolean;
    recruiter: boolean;
    progress: boolean;
    subject: string;
}